Add CucumberInfo/CucumberInfoBuilder
