+++ language-service/javascript/src/index.ts
@@ -1,3 +1,4 @@
 export * from './getGherkinSemanticTokens'
 export * from './getGherkinDiagnostics'
 export * from './getGherkinCompletionItems'
+export * from './CucumberInfo'
